refactor(dao): move testing related code to separate file
